# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Longevity Health Analytics Platform API",
    "description": "Longevity Health Analytics Platform API. Documentation-support software \u2014 not an autonomous medical decision-maker. Synthetic data only in development and tests.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/auth/dev-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Login",
        "operationId": "dev_login_api_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenancy/select-tenant": {
      "post": {
        "tags": [
          "tenancy"
        ],
        "summary": "Select Tenant",
        "operationId": "select_tenant_api_v1_tenancy_select_tenant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectTenantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_api_v1_patients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatientOut"
                  },
                  "type": "array",
                  "title": "Response List Patients Api V1 Patients Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Patient",
        "operationId": "create_patient_api_v1_patients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients/{patient_id}/observations": {
      "get": {
        "tags": [
          "observations"
        ],
        "summary": "Timeline",
        "operationId": "timeline_api_v1_patients__patient_id__observations_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelinePointOut"
                  },
                  "title": "Response Timeline Api V1 Patients  Patient Id  Observations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patients/{patient_id}/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Report",
        "operationId": "generate_report_api_v1_patients__patient_id__reports_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Generate Report Api V1 Patients  Patient Id  Reports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/{report_id}/approve": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Approve Report",
        "operationId": "approve_report_api_v1_reports__report_id__approve_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Report Api V1 Reports  Report Id  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/release": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Release Report",
        "operationId": "release_report_api_v1_reports__report_id__release_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patient-view": {
      "get": {
        "tags": [
          "patient-view"
        ],
        "summary": "Patient View",
        "operationId": "patient_view_api_v1_patient_view_get",
        "parameters": [
          {
            "name": "tenant",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patient View Api V1 Patient View Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePatientRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Display Name"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "CreatePatientRequest"
      },
      "DevLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "DevLoginRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "tenants": {
            "items": {
              "$ref": "#/components/schemas/TenantMembershipOut"
            },
            "type": "array",
            "title": "Tenants"
          }
        },
        "type": "object",
        "required": [
          "user",
          "tenants"
        ],
        "title": "LoginResponse"
      },
...
      "PatientOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "external_ref",
          "date_of_birth",
          "is_synthetic"
        ],
        "title": "PatientOut"
      },
      "ReleaseOut": {
        "properties": {
          "report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Report Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "patient_access_token": {
            "type": "string",
            "title": "Patient Access Token"
          },
          "patient_view_url": {
            "type": "string",
            "title": "Patient View Url"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "status",
          "patient_access_token",
          "patient_view_url"
        ],
        "title": "ReleaseOut"
      },
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "clinician",
          "assistant"
        ],
        "title": "Role",
        "description": "Tenant staff roles. Patients are not staff and have no Membership."
      },
      "SelectTenantRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "SelectTenantRequest"
      },
      "SelectTenantResponse": {
        "properties": {
          "active_tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Active Tenant Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "active_tenant_id",
          "role"
        ],
        "title": "SelectTenantResponse"
      },
...
      "TenantMembershipOut": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "tenant_name": {
            "type": "string",
            "title": "Tenant Name"
          },
          "tenant_slug": {
            "type": "string",
            "title": "Tenant Slug"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "tenant_name",
          "tenant_slug",
          "is_synthetic",
          "role"
        ],
        "title": "TenantMembershipOut"
      },
      "TimelinePointOut": {
        "properties": {
          "observation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Observation Id"
          },
          "metric_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Code"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "reference_low": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Low"
          },
          "reference_high": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference High"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "observed_at_is_date_only": {
            "type": "boolean",
            "title": "Observed At Is Date Only"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "delta_vs_previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Vs Previous"
          },
          "previous_observation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Observation Id"
          },
          "kpi_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kpi Code"
          },
          "kpi_primary_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kpi Primary Domain"
          },
          "kpi_secondary_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kpi Secondary Domains"
          },
          "comparability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparability"
          },
          "comparability_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparability Reason"
          }
        },
        "type": "object",
        "required": [
          "observation_id",
          "metric_code",
          "original_name",
          "value",
          "unit",
          "reference_low",
          "reference_high",
          "observed_at",
          "observed_at_is_date_only",
          "review_status",
          "delta_vs_previous",
          "previous_observation_id",
          "kpi_code",
          "kpi_primary_domain",
          "kpi_secondary_domains",
          "comparability",
          "comparability_reason"
        ],
        "title": "TimelinePointOut"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}